edit repository data
