+++ app/src/main/java/com/depromeet/android/childcare/model/response/GetExpendituresResponse.kt
@@ -1,6 +1,9 @@
 package com.depromeet.android.childcare.model.response
 
 import com.depromeet.android.childcare.model.PaymentType
+import com.depromeet.android.childcare.model.Record
+import com.depromeet.android.childcare.model.RecordType
+import com.depromeet.android.childcare.model.User
 
 data class GetExpendituresResponse(
     val `data`: List<Data>,
@@ -29,4 +32,19 @@ data class GetExpendituresResponse(
         val profileImageUrl: String,
         val status: String
     )
+
+    fun toRecords() = data.map {
+        Record(
+            it.id,
+            User(it.member.id, it.member.profileImageUrl, it.member.name, it.member.connectionCode),
+            RecordType.PAYMENT,
+            it.expendedAt,
+            it.title,
+            it.amountOfMoney,
+            it.category,
+            it.paymentMethod,
+            it.imageUrl,
+            it.description
+        )
+    }
 }
