+++ app/src/main/java/com/depromeet/android/childcare/model/response/MyInfoResponse.kt
@@ -1,5 +1,7 @@
 package com.depromeet.android.childcare.model.response
 
+import com.depromeet.android.childcare.model.User
+
 data class MyInfoResponse(
     val `data`: Data,
     val message: String,
@@ -16,5 +18,21 @@ data class MyInfoResponse(
 
     enum class Status {
         SOLO, COUPLE
+    }
+
+    fun meToUser() = User(
+        data.id,
+        data.profileImageUrl,
+        data.name,
+        data.connectionCode
+    )
+
+    fun spouseToUser() = data.spouseName?.let {
+        User(
+            -1,
+            null,
+            it,
+            ""
+        )
     }
 }
